+++ tests/test_main.py
@@ -12,6 +12,7 @@
 import errno
 import os
 import re
+from inspect import cleandoc
 from pathlib import Path
 from typing import Optional
 from unittest.mock import create_autospec
@@ -92,6 +93,104 @@ def test_lint_submodule_included_fail(submodule_repository, stringio):
 
     assert result == 1
     assert ":-(" in stringio.getvalue()
+
+
+def test_lint_meson_subprojects(fake_repository, stringio):
+    """Verify that subprojects are ignored."""
+    (fake_repository / "meson.build").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: CC0-1.0
+            """
+        ).replace("spdx", "SPDX")
+    )
+    subprojects_dir = fake_repository / "subprojects"
+    subprojects_dir.mkdir()
+    libfoo_dir = subprojects_dir / "libfoo"
+    libfoo_dir.mkdir()
+    # ./subprojects/foo.wrap has license and linter succeeds
+    (subprojects_dir / "foo.wrap").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: CC0-1.0
+            """
+        ).replace("spdx", "SPDX")
+    )
+    # ./subprojects/libfoo/foo.c misses license but is ignored
+    (libfoo_dir / "foo.c").write_text("foo")
+    result = main(["lint"], out=stringio)
+
+    assert result == 0
+    assert ":-)" in stringio.getvalue()
+
+
+def test_lint_meson_subprojects_fail(fake_repository, stringio):
+    """Verify that files in './subprojects' are not ignored."""
+    (fake_repository / "meson.build").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: CC0-1.0
+            """
+        ).replace("spdx", "SPDX")
+    )
+    subprojects_dir = fake_repository / "subprojects"
+    subprojects_dir.mkdir()
+    # ./subprojects/foo.wrap misses license and linter fails
+    (subprojects_dir / "foo.wrap").write_text("foo")
+    result = main(["lint"], out=stringio)
+
+    assert result == 1
+    assert ":-(" in stringio.getvalue()
+
+
+def test_lint_meson_subprojects_included_fail(fake_repository, stringio):
+    """When Meson subprojects are included, fail on errors."""
+    (fake_repository / "meson.build").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: CC0-1.0
+            """
+        ).replace("spdx", "SPDX")
+    )
+    libfoo_dir = fake_repository / "subprojects/libfoo"
+    libfoo_dir.mkdir(parents=True)
+    # ./subprojects/libfoo/foo.c misses license and linter fails
+    (libfoo_dir / "foo.c").write_text("foo")
+    result = main(["--include-meson-subprojects", "lint"], out=stringio)
+
+    assert result == 1
+    assert ":-(" in stringio.getvalue()
+
+
+def test_lint_meson_subprojects_included(fake_repository, stringio):
+    """Successfully lint when Meson subprojects are included."""
+    (fake_repository / "meson.build").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: CC0-1.0
+            """
+        ).replace("spdx", "SPDX")
+    )
+    libfoo_dir = fake_repository / "subprojects/libfoo"
+    libfoo_dir.mkdir(parents=True)
+    # ./subprojects/libfoo/foo.c has license and linter succeeds
+    (libfoo_dir / "foo.c").write_text(
+        cleandoc(
+            """
+            spdx-FileCopyrightText: 2022 Jane Doe
+            spdx-License-Identifier: GPL-3.0-or-later
+            """
+        ).replace("spdx", "SPDX")
+    )
+    result = main(["--include-meson-subprojects", "lint"], out=stringio)
+
+    assert result == 0
+    assert ":-)" in stringio.getvalue()
 
 
 def test_lint_fail(fake_repository, stringio):